requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib[bcrypt]
bcrypt==4.3.0
python-jose[cryptography]
PyJWT
itsdangerous
asyncpg
redis
celery[redis]
flower
python-dateutil
email-validator
python-multipart
fastapi-mail
alembic
pydantic-settings